+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 - fix webxdc: allow `self` and `blob:` in `connect-src` in CSP
+- indentation in update device message
 
 <a id="1_34_0"></a>