+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 ### Fixed
+- fix webxdc: allow `self` and `blob:` in `connect-src` in CSP
 - indentation in update device message
 
 <a id="1_34_0"></a>