+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - show jump down button earlier when scrolling up
+- make font-style of info messages consistent with iOS and Android #3034
 
 ### Fixed
 - fix webxdc: allow `self` and `blob:` in `connect-src` in CSP