+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - indentation in update device message
 - fix "no messages" blinking up for a second when loading a chat
 - hide unread count on jump down button if you are at the bottom, fixes #3033
+- fix removing group avatar image #3038
 
 <a id="1_34_0"></a>