+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - make font-style of info messages consistent with iOS and Android #3034
 - Bump `decode-uri-component` from `0.2.0` to `0.2.2`.
 - move all verified icons to the end chat/contact name (previously some were still on the avatar image)
+- set `ITSAppUsesNonExemptEncryption` to `false` for mac
 
 ### Fixed
 - fix webxdc: allow `self` and `blob:` in `connect-src` in CSP